Move testRemovals function to import to avoid hoisting bug
